+++ CHANGELOG.md
@@ -17,6 +17,16 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 ### Added
+- most task properties can now also be defined as CLI arguments (#23)
+
+### Changed
+- Removed dependency on java-plugin (#19)
+
+### Fixed
+- JMX-File now is really optional for gui tasks (#20)
+
+## [2.0.0]
+### Added
 - Support for more JMeter cli properties (#12)
 - mandatory JMeter configurations are now handled by dependency on respective config-jar (#15)